# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "E.Practice API",
    "description": "Modulo case management per studi professionali — backend FastAPI V0 (in-memory). Differenziatore: vista albero della pratica con eventi ad-hoc agganciati alle fasi.",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/clients/{client_id}/practices": {
      "get": {
        "tags": [
          "clients"
        ],
        "summary": "List Client Practices",
        "description": "Lista pratiche del cliente (per scheda cliente).",
        "operationId": "list_client_practices_api_clients__client_id__practices_get",
        "parameters": [
          {
            "name": "client_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Client Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Practice"
                  },
                  "title": "Response List Client Practices Api Clients  Client Id  Practices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/practices": {
      "get": {
        "tags": [
          "practices"
        ],
        "summary": "List Practices",
        "operationId": "list_practices_api_practices_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Ricerca su title, code",
              "title": "Q"
            },
            "description": "Ricerca su title, code"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "aperta",
                    "in_corso",
                    "in_attesa",
                    "sospesa",
                    "chiusa",
                    "archiviata"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "bassa",
                    "media",
                    "alta"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "name": "responsible_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Responsible Id"
            }
          },
          {
            "name": "client_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Client Id"
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Practice_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "practices"
        ],
        "summary": "Create Practice",
        "description": "Wizard creazione pratica: genera Practice + instanzia fasi dal template.\n\nOpzionale: se `create_default_reminders=True`, crea anche un Reminder per\nogni fase (days_before=2). Se `label_ids` non vuoto, aggancia le etichette.",
        "operationId": "create_practice_api_practices_post",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePracticeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePracticeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/practices/{practice_id}": {
      "get": {
        "tags": [
          "practices"
        ],
        "summary": "Get Practice Detail",
        "description": "Detail completo con joined names — una sola GET per la pagina.",
        "operationId": "get_practice_detail_api_practices__practice_id__get",
        "parameters": [
          {
            "name": "practice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Practice Id"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/practices/{practice_id}/phases": {
      "get": {
        "tags": [
          "practices"
        ],
        "summary": "List Practice Phases",
        "operationId": "list_practice_phases_api_practices__practice_id__phases_get",
        "parameters": [
          {
            "name": "practice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Practice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PracticePhase"
                  },
                  "title": "Response List Practice Phases Api Practices  Practice Id  Phases Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/practices/{practice_id}/events": {
      "get": {
        "tags": [
          "practices"
        ],
        "summary": "List Practice Events",
        "operationId": "list_practice_events_api_practices__practice_id__events_get",
        "parameters": [
          {
            "name": "practice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Practice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PracticeEvent"
                  },
                  "title": "Response List Practice Events Api Practices  Practice Id  Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/practices/{practice_id}/archive": {
      "post": {
        "tags": [
          "practices"
        ],
        "summary": "Archive Practice",
        "operationId": "archive_practice_api_practices__practice_id__archive_post",
        "parameters": [
          {
            "name": "practice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Practice Id"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Practice"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/practices/{practice_id}/labels": {
      "get": {
        "tags": [
          "practices"
        ],
        "summary": "List Practice Labels",
        "operationId": "list_practice_labels_api_practices__practice_id__labels_get",
        "parameters": [
          {
            "name": "practice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Practice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LabelSummary"
                  },
                  "title": "Response List Practice Labels Api Practices  Practice Id  Labels Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePracticeResponse": {
        "properties": {
          "practice_id": {
            "type": "string",
            "format": "uuid",
            "title": "Practice Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "phase_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Phase Ids"
          }
        },
        "type": "object",
        "required": [
          "practice_id",
          "code",
          "phase_ids"
        ],
        "title": "CreatePracticeResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LabelSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "color": {
            "type": "string",
            "title": "Color"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "color"
        ],
        "title": "LabelSummary"
      },
...
      "Page_Practice_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Practice"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "Page[Practice]"
      },
...
      "Practice": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "client_id": {
            "type": "string",
            "format": "uuid",
            "title": "Client Id"
          },
          "client_token": {
            "type": "string",
            "title": "Client Token"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "responsible_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsible Id"
          },
          "apertura": {
            "type": "string",
            "format": "date",
            "title": "Apertura"
          },
          "scadenza": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scadenza"
          },
          "priority": {
            "type": "string",
            "enum": [
              "bassa",
              "media",
              "alta"
            ],
            "title": "Priority",
            "default": "media"
          },
          "status": {
            "type": "string",
            "enum": [
              "aperta",
              "in_corso",
              "in_attesa",
              "sospesa",
              "chiusa",
              "archiviata"
            ],
            "title": "Status"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "code",
          "title",
          "client_id",
          "client_token",
          "category_id",
          "apertura",
          "status",
          "created_by",
          "id",
          "created_at"
        ],
        "title": "Practice"
      },
...
    }
  }
}